+++ .gitignore
@@ -30,6 +30,8 @@
 *.exe
 *.out
 *.app
+*.pdb
+*.ilk
 /bin/documentation
 /bin/rom
 /bin/disk
@@ -37,3 +39,23 @@ granite.depend
 bin/config.txt
 *.img
 *.raw
+
+# Build Directories
+/build/
+
+# In-place Compilation
+.ninja_*
+build.ninja
+CMakeLists.txt.user
+CMakeCache.txt
+CMakeFiles
+CMakeScripts
+Testing
+Makefile
+cmake_install.cmake
+install_manifest.txt
+compile_commands.json
+CTestTestfile.cmake
+_deps
+CMakeUserPresets.json
+deps
