+++ include/diskette.h
