+++ include/palette.h
